% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disag.R
\docType{package}
\name{agouti}
\alias{agouti}
\title{Disaggregation regression classes and basic models.}
\description{
Utility functions, classes and data visualisation for aggregate outcomes
models.
}
\author{
Tim CD Lucas
}
