% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{agouti_summary}
\alias{agouti_summary}
\title{Summarise a aggregated output dataset.}
\usage{
agouti_summary(x, ID = ID)
}
\arguments{
\item{x}{A data frame suitable for use as aggregate output data.}

\item{ID}{The column to use as the group ID column. Unquoted value.}
}
\description{
Summarise a aggregated output dataset.
}
\examples{
data(madagascar_malaria)
agouti_summary(madagascar_malaria)
}
