% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agouti_fit.R
\name{agoutiGLM}
\alias{agoutiGLM}
\title{Fit an aggregated output GLM}
\usage{
agoutiGLM(formula, data, ID, inner_link, outer_link, family, weights, ...)
}
\arguments{
\item{formula}{A formula for the y and x variables.}

\item{data}{A data frame suitable for aggregate outputs models.}

\item{ID}{The column to use as the group ID column. Unquoted value.}

\item{inner_link}{A character defining the inner link function.}

\item{outer_link}{A character defining the outer link function.}

\item{family}{The error distribution defined with a family object or character.}

\item{weights}{Which column to use as within group weights such as 
population or time.}

\item{...}{Further parameters. Currently unused.

data(madagascar_malaria)
agoutiGLM(case_rate ~ LSTmean, data = madagascar_malaria, ID = ID, 
          weights = pop)}
}
\description{
This function has flexibility for link functions both inside and outside the
aggregation step as well as a variety of error models. The models are fitted
with maximum likelihood and confidence intervals are given using the normal
approximation to the likelihood surface.
}
