% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_disag.R
\name{as_disag.data.frame}
\alias{as_disag.data.frame}
\title{Checks if a given dataframe is in the correct format for agouti functions}
\usage{
\method{as_disag}{data.frame}(data, data2 = NULL, response_var, ...)
}
\arguments{
\item{data}{A dataframe containing a outcome variable and ID variable named ID}

\item{data2}{A dataframe default is NULL in which case we check the data is already in the correct format. If not null,
the second data frame should contain predictor variables at a different resolution to the outcome but with the same linking ID variable}

\item{response_var}{The name of the variable containing the response data}

\item{...}{any other arguments}
}
\value{
\code{as_disag.data.frame} returns a dataframe that can be used with all agouti functions.
}
\description{
Checks if a given dataframe is in the correct format for agouti functions
}
\examples{
\dontrun{
disag_data <- as_disag(data=df_outcome, data2=df_predictor, outcome="Death")
}
data(madagascar_malaria)
disag_data <- as_disag(data=madagascar_malaria,response_var="case_rate")
}
