% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_disag.R
\name{as_disag.POSIXt}
\alias{as_disag.POSIXt}
\title{Create a data.frame suitable for disaggregation regression modelling from
a temporal dataframe object.}
\usage{
\method{as_disag}{POSIXt}(
  data,
  time_group = "\%Y\%m\%d",
  ID = "add",
  response_df,
  outcome = "Death",
  ...
)
}
\arguments{
\item{data}{a column of a dataframe that is in datetime format}

\item{time_group}{what is the format of the datetime variable at the level of the response in date-time format}

\item{ID}{Either the name of the grouping variable or "add" which means an ID variable will be added 1:nrow()}

\item{response_df}{a dataframe containing at least data,ID and outcome}

\item{outcome}{A character string naming the variable which stores the outcome data}

\item{...}{any other arguments}
}
\value{
\code{as_disag.POSIXt} returns a dataframe that can be used with all agouti functions
}
\description{
A series of lagged variables can be considered similar to disaggregation
regression if we think that the order of the lagged variables is unimportant.
For example, if we have millisecond resolution airpollution data relating to
daily health outcomes, it is unimportant whether an airpollution measurement
occured 2 miliseconds or 4 miliseconds before the health data cutoff time.
}
\section{More stuff}{
NA
}

\examples{
\dontrun{
data(mortality_temporal)
disag_data <- as_disag(data=mortality_temporal$Datetime,
response_df=mortality_temporal, outcome="Death")
}
}
