% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_disag.R
\name{as_disag}
\alias{as_disag}
\title{as_dsiag Generic function}
\usage{
as_disag(data, ...)
}
\arguments{
\item{data}{A dataframe containing at least an outcome variable and ID variable named ID}

\item{...}{any other arguments}
}
\description{
as_dsiag Generic function
}
