% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_disag.R
\name{as_disag.sf}
\alias{as_disag.sf}
\title{Create a data.frame suitable for disaggregation regression modelling from
a sf object.}
\usage{
\method{as_disag}{sf}(data, rstack, response_var = "response", ...)
}
\arguments{
\item{data}{An object of class sf containing polygons that covariates should be aggregated to and including the response variable and ID}

\item{rstack}{A raster stack of class SpatRaster to be summarised at ID level}

\item{response_var}{Name of the response variable found in data}

\item{...}{any other arguments}
}
\value{
\code{as_disag.sf} returns a dataframe that can be used with all agouti functions.
}
\description{
Create a data.frame suitable for disaggregation regression modelling from
a sf object.
}
\examples{
\dontrun{
disag_data <- as_disag(data=shapefile, rstack=rstack, response_var=Human.cases)
}
}
