% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_disag.R
\name{as_disag.SpatialPolygonsDataFrame}
\alias{as_disag.SpatialPolygonsDataFrame}
\title{Create a data.frame suitable for disaggregation regression modelling from
a spatial polygons dataframe.}
\usage{
\method{as_disag}{SpatialPolygonsDataFrame}(data, rstack, response_df, ...)
}
\arguments{
\item{data}{A spatial polygons data frame with a column named ID}

\item{rstack}{A raster stack of rasters to be summarised at ID level}

\item{response_df}{a dataframe containing the response data and an ID variable}

\item{...}{any other arguments}
}
\value{
\code{as_disag.SpatialPolygonsDataFrame} returns a dataframe that can be used with all agouti functions.
}
\description{
Create a data.frame suitable for disaggregation regression modelling from
a spatial polygons dataframe.
}
\examples{
\dontrun{
disag_data <- as_disag(data=shapefile, rstack=rstack, response=wnv)
}
}
