% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_disag.R
\name{as_disag.ts}
\alias{as_disag.ts}
\title{Create a data.frame suitable for disaggregation regression modelling from
a time series object.}
\usage{
\method{as_disag}{ts}(data, lags = 10, ID = "add", ...)
}
\arguments{
\item{data}{A vector of class time series representing a response e.g. in stock data the growth}

\item{lags}{An integer saying how many lags should be taken}

\item{ID}{Either the name of the grouping variable or "add" which means an ID variable will be added 1:nrow()}

\item{...}{any other arguments}
}
\value{
\code{as_disag.ts} returns a dataframe that can be used with all agouti functions.
}
\description{
An example time series object is daily stock price growth
}
\section{More stuff}{
NA
}

\examples{
data(stock_vector)
disag_data <- as_disag(data=stock_vector)
}
