% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_disag.R
\name{disag_from_temporal_dataframe}
\alias{disag_from_temporal_dataframe}
\title{Create a data.frame suitable for disaggregation regression modelling from
a time series object.}
\usage{
disag_from_temporal_dataframe(data, lags)
}
\arguments{
\item{data}{A time series data object}

\item{lags}{An integer list given which lags should be used.}
}
\value{
Nothing at the moment
}
\description{
A series of lagged variables can be considered similar to disaggregation 
regression if we think that the order of the lagged variables is unimportant.
For example, if we have millisecond resolution airpollution data relating to 
daily health outcomes, it is unimportant whether an airpollution measurement
occured 2 miliseconds or 4 miliseconds before the health data cutoff time.
}
\section{More stuff}{
NA
}

\examples{
disag_data <- disag_from_temporal_dataframe()
}
