% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_hist.R, R/plotting_functions.R
\name{heat_hist}
\alias{heat_hist}
\title{A grouped histogram plot.}
\usage{
heat_hist(x, ID, weights = 1, breaks = 50)

heat_hist(x, ID, weights = 1, breaks = 50)
}
\description{
A grouped histogram plot.

A grouped histogram plot.
}
