% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_multi_hist.R
\name{heat_multi_hist}
\alias{heat_multi_hist}
\title{A grouped histogram multi plot.}
\usage{
heat_multi_hist(y, ID, data, weights = 1, breaks = 50)
}
\arguments{
\item{y}{A character vector of the variable names to be plotted (first variable listed is how the data gets ordered).}

\item{ID}{A character or factor vector to be used as a grouping variable.}

\item{data}{A dataframe containing the variables to be plotted}

\item{weights}{A numeric vector giving the weights.}

\item{breaks}{How many breaks to use.}
}
\description{
A grouped histogram multi plot.
}
\examples{
data(madagascar_malaria)
heat_multi_hist(c("Elevation","EVI","LSTmean"), madagascar_malaria$ID, data=madagascar_malaria)
}
