% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agouti_fit.R
\name{loglike}
\alias{loglike}
\title{Calculate the log likelihood for each observation given a set of parameters.}
\usage{
loglike(beta, x, y, ID, inner, outer, likelihood_function, weights)
}
\arguments{
\item{beta}{The parameter vector}

\item{x}{covariate matrix}

\item{y}{observation vector.}

\item{ID}{The column to use as the group ID column. Unquoted value.}

\item{inner}{A character defining the inner link function.}

\item{outer}{A character defining the outer link function.}

\item{likelihood_function}{The likelihood (error) function given as a 
function such as dnorm.}

\item{weights}{Which column to use as within group weights such as 
population or time.}
}
\description{
Calculate the log likelihood for each observation given a set of parameters.
}
