% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_madagascar_malaria.R
\docType{data}
\name{madagascar_malaria}
\alias{madagascar_malaria}
\title{Madagascar Malaria data}
\format{
A data.frame with x observations and y variables:
\describe{
\item{ID}{An ID column for the areal data with}
\item{Temp, NVDI}{Temperature and vegeation index covariates}
}
}
\usage{
madagascar_malaria
}
\description{
This data includes areal counts of malaria cases in madagascar, and high 
resolution covariates and population counts.
The format is one row per high resolution pixel, with the values for the areal
data being repeated. And ID column unambiguously links the two scales.
}
\examples{
madagascar_malaria
}
\seealso{
The script to create this data set: \url{}
}
\keyword{datasets}
