% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mortality_temporal.R
\docType{data}
\name{mortality_temporal}
\alias{mortality_temporal}
\title{Temporal respiratory mortality data
From nidhi}
\format{
A data.frame with x observations and y variables:
\describe{
\item{Datetime}{An column of class datetime}
\item{Temp}{Hourly temperature data}
\item{Death}{Daily deaths}

}
}
\usage{
mortality_temporal
}
\description{
Temporal respiratory mortality data
From nidhi
}
\examples{
data(mortality_temporal)
}
\seealso{
The script to create this data set: \url{}
}
\keyword{datasets}
