% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stock_vector.R
\docType{data}
\name{stock_vector}
\alias{stock_vector}
\title{Stock vector}
\format{
A ts vector with x observations:
\describe{
\item{growth}{A vector containing daily stock growths}
}
}
\usage{
stock_vector
}
\description{
This data contains stock data for the companys in the S&P500.
The response variable, growth, is the a stocks closing price, divded by it's
closing price the previous day, minus 1 (i.e. daily percentage change)
}
\examples{
stock_vector
}
\seealso{
The script to create this data set: \url{}
}
\keyword{datasets}
