% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_summary_plot.R
\name{group_summary_plot}
\alias{group_summary_plot}
\title{A function that summarises the groups within an aggregate output dataset, 
by taking quantiles, and then plots these quantiles against the response data.}
\usage{
group_summary_plot(formula, data, ID = ID, weights = weights)
}
\description{
A function that summarises the groups within an aggregate output dataset, 
by taking quantiles, and then plots these quantiles against the response data.
}
\examples{

data(madagascar_malaria)
group_summary_plot(case_rate ~ LSTmean, data = madagascar_malaria, ID = ID, weights = pop)
}
