% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_hist.R
\name{heat_hist}
\alias{heat_hist}
\title{A grouped histogram plot.}
\usage{
heat_hist(x, ID, weights = 1, breaks = 50)
}
\arguments{
\item{x}{A numeric vector to be plotted as histograms.}

\item{ID}{A character or factor vector to be used as a grouping variable.}

\item{weights}{A numeric vector giving the weights.}
}
\description{
A grouped histogram plot.
}
\examples{
data(madagascar_malaria)
heat_hist(madagascar_malaria$EVI, madagascar_malaria$ID)
}
