% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_plot.R
\name{link_plot}
\alias{link_plot}
\title{A very basic plotting for aggregate data that plots a linked scatter plot.}
\usage{
link_plot(formula, data, ID = ID, weights = weights)
}
\arguments{
\item{ID}{A character or factor vector to be used as a grouping variable.}

\item{weights}{A numeric vector giving the weights.}
}
\description{
This is a scatter plot which data from the aggregate unit "linked" by 
a line and colour coded to match.
}
\examples{
data(madagascar_malaria)
link_plot(case_rate ~ LSTmean, data = madagascar_malaria, ID = ID, weights = pop)
}
