% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_plot.R
\name{link_plot}
\alias{link_plot}
\title{A very basic plotting for aggregate data that plots a linked scatter plot.}
\usage{
link_plot(formula, data, ID = ID, weights = weights)
}
\arguments{
\item{formula}{A formula for the y and x variables.}

\item{data}{A data frame suitable for aggregate outputs models.}

\item{ID}{The column to use as the group ID column. Unquoted value.}

\item{weights}{Which column to use as within group weights such as 
  population or time.

data(madagascar_malaria)
group_summary_plot(case_rate ~ LSTmean, data = madagascar_malaria, ID = ID, weights = pop)}
}
\description{
This is a scatter plot which data from the aggregate unit "linked" by 
a line and colour coded to match.
}
\examples{
data(madagascar_malaria)
link_plot(case_rate ~ LSTmean, data = madagascar_malaria, ID = ID, weights = pop)
}
