% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_small_multiple.R
\name{thresh_sm}
\alias{thresh_sm}
\title{A plot that summarised the relationship between a variable and an aggregated
output that depends on it.}
\usage{
thresh_sm(
  formula,
  data,
  ID = ID,
  weights = weights,
  small_mult = 25,
  lower_percentile = 0.05,
  upper_percentile = 0.95
)
}
\description{
The plot is created by calculating, within each group, the proportion of 
observations that are above a threshold. A scatter plot with a linear model
is plotted of the relationship between this proportion and the aggregated output.
The threshold is varied and small multiples of the different values are plotted.
}
